requirements.txt
tornado
pillow
opencv-python
pymongo
wtforms
werkzeug
face_recognition
